Remove base Parser class
In Snorkel, with CoreNLP we needed a connection interface for the Java
server. With spaCy, it's just a function call.
