+++ fonduer/parser/__init__.py
@@ -1,13 +1,12 @@
 from __future__ import absolute_import
 
-from fonduer.parser.async_parser import OmniParser
 from fonduer.parser.doc_preprocessors import (CSVPathsPreprocessor,
                                               DocPreprocessor,
                                               HTMLPreprocessor,
                                               TextDocPreprocessor,
                                               TSVDocPreprocessor,
                                               XMLMultiDocPreprocessor)
-from fonduer.parser.parser import Parser, ParserConnection
+from fonduer.parser.parser import OmniParser
 from fonduer.parser.rule_parser import (RegexTokenizer, SpacyTokenizer,
                                         Tokenizer)
 from fonduer.parser.spacy_parser import Spacy
